Responsibility-splitting clean up; http caching
Extract a couple of new classes from Podcast to make it into an Entity,
with a Repository-pattern class for finding / co-ordinated fetching data
for this, along with a PodcastBuilder to orchestrate pulling together
the stuff needed for a Podcast instance.

This was initially kicked off by the need to have access to some data
about a podcast that was harder to get - to add an Etag and last
modified time to the RSS feed.

Once done I was able to pretty easily add the ETag and Last-Modified /
Expires headers I wanted.